+++ TypeOfShape.Application/TypeOfShape/ITypeOfShapeService.cs
@@ -3,5 +3,5 @@
 public interface ITypeOfShapeService<TShapeType>
     where TShapeType : Enum
 {
-    ErrorOr<TShapeType> Handle(float[] sides);
+    ErrorOr<TShapeType> Handle(double[] sides);
 }
